+++ data-raw/DataComputing.R
@@ -36,3 +36,9 @@ Encoding(WorldCities$name) <- "latin1"
 WorldCities$name <- iconv(WorldCities$name, from = "latin1", to = "UTF-8")
 save(WorldCities, file = "data/WorldCities.rda", compress = "xz")
 
+# NCI60 -- 5 MB file so put it in a different place
+download.file("https://github.com/DataComputing/DataComputing/raw/master/data/NCI60.rda", "data-raw/NCI60.rda")
+load("data-raw/NCI60.rda")
+NCI60_tiny <- head(NCI60)
+save(NCI60_tiny, file = "data/NCI60_tiny.rda", compress = "xz")
+cat(create_roxygen(parse_file("https://raw.githubusercontent.com/DataComputing/DataComputing/master/man/NCI60.Rd")), sep = "\n")