+++ NEWS.md
@@ -1,4 +1,4 @@
-# mdsr 0.1.4
+mdsr 0.1.4
 
 * Added CRAN badge after release.
 * Added `etl_NCI60()` function